+++ Modules/Microsoft365DSC/Examples/Resources/AADGroupsSettings/1-ConfigureGroupsSettings.ps1
@@ -8,7 +8,7 @@ Configuration Example
     param(
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -25,7 +25,7 @@ Configuration Example
             GuestUsageGuidelinesUrl       = "https://contoso.com/guestusage"
             UsageGuidelinesUrl            = "https://contoso.com/usage"
             Ensure                        = "Present"
-            Credential                    = $credsGlobalAdmin
+            Credential                    = $Credscredential
         }
     }
 }