+++ Modules/Microsoft365DSC/Examples/Resources/IntuneAppConfigurationPolicy/1-RemoveAppConfigurationPolicy.ps1
@@ -7,7 +7,7 @@ Configuration Example
     param(
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -18,7 +18,7 @@ Configuration Example
             DisplayName = 'ContosoOld'
             Description = 'Old Contoso Policy'
             Ensure      = 'Absent'
-            Credential  = $credsGlobalAdmin
+            Credential  = $Credscredential
         }
     }
 }