+++ Modules/Microsoft365DSC/Examples/Resources/IntuneDeviceConfigurationPolicyAndroidDeviceOwner/1-ConfigureIntuneDeviceConfigurationPolicyAndroidDeviceOwner.ps1
@@ -8,7 +8,7 @@ Configuration Example
     param(
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -59,7 +59,7 @@ Configuration Example
             WorkProfilePasswordRequiredType       = 'deviceDefault'
             WorkProfilePasswordRequireUnlock      = 'deviceDefault'
             Ensure                                = 'Present'
-            Credential                            = $credsGlobalAdmin
+            Credential                            = $Credscredential
         }
     }
 }