+++ Modules/Microsoft365DSC/Examples/Resources/IntuneDeviceConfigurationPolicyAndroidOpenSourceProject/1-ConfigureIntuneDeviceConfigurationPolicyAndroidOpenSourceProject.ps1
@@ -8,7 +8,7 @@ Configuration Example
     param(
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -30,7 +30,7 @@ Configuration Example
             ScreenCaptureBlocked      = $True
             StorageBlockExternalMedia = $True
             Ensure                    = 'Present'
-            Credential                = $credsGlobalAdmin
+            Credential                = $Credscredential
         }
     }
 }