+++ Modules/Microsoft365DSC/Examples/Resources/IntuneRoleAssignment/1-IntuneRoleAssignment.ps1
@@ -7,7 +7,7 @@ Configuration Example
     param(
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -26,7 +26,7 @@ Configuration Example
             RoleDefinition             = '2d00d0fd-45e9-4166-904f-b76ac5eed2c7'
             RoleDefinitionDisplayName  = 'This is my role'
             Ensure                     = 'Present'
-            Credential                 = $credsGlobalAdmin
+            Credential                 = $Credscredential
         }
     }
 }