+++ Modules/Microsoft365DSC/Examples/Resources/SCComplianceTag/1-AddingNewComplianceTag.ps1
@@ -8,7 +8,7 @@ Configuration Example
     param(
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -31,7 +31,7 @@ Configuration Example
                 FilePlanPropertySubcategory = "DemoSub"
             }
             Ensure            = "Present"
-            Credential        = $credsGlobalAdmin
+            Credential        = $Credscredential
         }
     }
 }