+++ Modules/Microsoft365DSC/Examples/Resources/SCDeviceConditionalAccessPolicy/1-ConfigureDeviceConditionalAccessPolicy.ps1
@@ -9,7 +9,7 @@ Configuration Example
     (
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -21,7 +21,7 @@ Configuration Example
             Comment              = "Device Conditional Access Policy for Human Resources department"
             Enabled              = $True
             Ensure               = "Present"
-            Credential           = $credsGlobalAdmin
+            Credential           = $Credscredential
         }
     }
 }