+++ Modules/Microsoft365DSC/Examples/Resources/SCFilePlanPropertyAuthority/1-AddingNewFilePlanPropertyCategory.ps1
@@ -9,7 +9,7 @@ Configuration Example
     (
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
 
     Import-DscResource -ModuleName Microsoft365DSC
@@ -20,7 +20,7 @@ Configuration Example
         {
             Name       = 'My Authority'
             Ensure     = 'Present'
-            Credential = $credsGlobalAdmin
+            Credential = $Credscredential
         }
     }
 }