+++ Modules/Microsoft365DSC/Examples/Resources/SCSecurityFilter/1-AddingNewSecurityFilter.ps1
@@ -8,7 +8,7 @@ Configuration Example
     param(
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -23,7 +23,7 @@ Configuration Example
             Filters         = @("Mailbox_CountryCode -eq '124'")
             Region          = "AUS"
             Ensure          = "Present"
-            Credential      = $credsGlobalAdmin
+            Credential      = $Credscredential
         }
     }
 }