+++ Modules/Microsoft365DSC/Examples/Resources/SPOPropertyBag/1-AddNewPropertyBagKey.ps1
@@ -8,7 +8,7 @@ Configuration Example
     param(
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -20,7 +20,7 @@ Configuration Example
             Key        = "MyKey"
             Value      = "MyValue"
             Ensure     = "Present"
-            Credential = $credsGlobalAdmin
+            Credential = $Credscredential
         }
     }
 }