+++ Modules/Microsoft365DSC/Examples/Resources/TeamsClientConfiguration/1-ConfigureTeamsClient.ps1
@@ -8,7 +8,7 @@ Configuration Example
     param(
         [Parameter(Mandatory = $true)]
         [PSCredential]
-        $credsGlobalAdmin
+        $Credscredential
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
@@ -29,7 +29,7 @@ Configuration Example
             ContentPin                       = "RequiredOutsideScheduleMeeting"
             Identity                         = "Global"
             ResourceAccountContentAccess     = "NoAccess"
-            Credential                       = $credsGlobalAdmin
+            Credential                       = $Credscredential
         }
     }
 }