+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.AADConditionalAccessPolicy.Tests.ps1
@@ -67,7 +67,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     ExcludeExternalTenantsMembers        = @()
                     ExcludeExternalTenantsMembershipKind = 'all'
                     ExcludeGuestOrExternalUserTypes      = @('internalGuest', 'b2bCollaborationMember')
-                    Credential                           = $Credsglobaladmin
+                    Credential                           = $Credscredential
                     GrantControlOperator                 = 'AND'
                     Id                                   = 'bcc0cf19-ee89-46f0-8e12-4b89123ee6f9'
                     IncludeApplications                  = @('All')
@@ -162,7 +162,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     ExcludeExternalTenantsMembers            = @()
                     ExcludeExternalTenantsMembershipKind     = 'all'
                     ExcludeGuestOrExternalUserTypes          = @('internalGuest', 'b2bCollaborationMember')
-                    Credential                               = $Credsglobaladmin
+                    Credential                               = $Credscredential
                     GrantControlOperator                     = 'AND'
                     Id                                       = 'bcc0cf19-ee89-46f0-8e12-4b89123ee6f9'
                     IncludeApplications                      = @('All')
@@ -316,7 +316,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     ExcludeExternalTenantsMembers            = @()
                     ExcludeExternalTenantsMembershipKind     = 'all'
                     ExcludeGuestOrExternalUserTypes          = @('internalGuest', 'b2bCollaborationMember')
-                    Credential                               = $Credsglobaladmin
+                    Credential                               = $Credscredential
                     GrantControlOperator                     = 'AND'
                     Id                                       = 'bcc0cf19-ee89-46f0-8e12-4b89123ee6f9'
                     IncludeApplications                      = @('All')
@@ -484,7 +484,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     ExcludePlatforms                         = @('Windows', 'WindowsPhone', 'MacOS')
                     ExcludeRoles                             = @('Compliance Administrator')
                     ExcludeUsers                             = 'alexw@contoso.com'
-                    Credential                               = $Credsglobaladmin
+                    Credential                               = $Credscredential
                     GrantControlOperator                     = 'AND'
                     Id                                       = 'bcc0cf19-ee89-46f0-8e12-4b89123ee6f9'
                     IncludeApplications                      = @('All')