+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change log for Microsoft365DSC
 
+# 1.23.1213.1
+
+* IntuneEndpointDetectionAndResponsePolicyWindows10
+  * Fix issue with assignments
+    FIXES [#3904](https://github.com/microsoft/Microsoft365DSC/issues/3904)
+* IntuneAntivirusPolicyWindows10SettingCatalog
+  * Fix issue with Set-TargetResource when retrieving a policy from displayName
+    FIXES [#4003](https://github.com/microsoft/Microsoft365DSC/issues/4003)
+* IntuneWindowsUpdateForBusinessRingUpdateProfileWindows10
+  * Fix parameter name in assignment cmdlet
+    FIXES [#4007](https://github.com/microsoft/Microsoft365DSC/issues/4007)
+* DEPENDENCIES
+  * Updated Microsoft.Graph to version 2.11.1.
+  * Updated MSCloudLoginAssistant to version 1.1.3.
+
 # 1.23.1206.1
 
 * IntuneAntivirusPolicyWindows10SettingCatalog